moved expenses.dart into widgets folder
