+++ lib/widgets/expenses.dart
@@ -1,0 +1,37 @@
+import 'package:expense_tracker_app/widgets/expenses_list/expenses_list.dart';
+import 'package:flutter/material.dart';
+import 'package:expense_tracker_app/models/expense.dart';
+
+class Expenses extends StatefulWidget {
+  const Expenses({super.key});
+
+  @override
+  State<Expenses> createState() => _ExpenseState();
+}
+
+class _ExpenseState extends State<Expenses> {
+  final List<Expense> _registeredExpense = [
+    Expense(
+      title: 'Flutter Course',
+      amount: 19.99,
+      date: DateTime.now(),
+      category: Category.work,
+    ),
+    Expense(
+      title: 'Cinema',
+      amount: 15.69,
+      date: DateTime.now(),
+      category: Category.leisure,
+    )
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Expanded(child: ExpensesList(expenses: _registeredExpense),
+        )
+      ],
+    );
+  }
+}
